+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 
 # CHANGELOG
 
+## v0.5.10 2020-6-3
+
+* Add a sync hook `validate` in the `Manager` trait to *quickly* determine connections are still valid when check-in
+
+## v0.5.9 2020-6-3
+
+* Fix get timeout.
+
+## v0.5.8(yanked) 2020-6-1
+
+* Add a hook to determine connections are valid when check-in.
+
 ## v0.5.7 2020-03-28
 
 * Fix PANIC: overflow when subtracting duration from instant.